+++ codegen-client/src/main/kotlin/software/amazon/smithy/rust/codegen/client/smithy/protocols/HttpBoundProtocolGenerator.kt
@@ -82,13 +82,12 @@ class HttpBoundProtocolTraitImplGenerator(
     )
     private val orchestratorCodegenScope by lazy {
         arrayOf(
-            "ConfigBag" to CargoDependency.smithyRuntime(runtimeConfig).toType().resolve("ConfigBag"),
-            "Error" to CargoDependency.smithyRuntimeApi(runtimeConfig).toType().resolve("interceptors::context::Error"),
-            "HttpResponse" to CargoDependency.smithyRuntime(runtimeConfig).toType().resolve("HttpResponse"),
+            "Error" to CargoDependency.smithyRuntimeApi(runtimeConfig).toType().resolve("client::interceptors::context::Error"),
+            "HttpResponse" to CargoDependency.smithyRuntime(runtimeConfig).toType().resolve("client::orchestrator::HttpResponse"),
             "Instrument" to CargoDependency.Tracing.toType().resolve("Instrument"),
-            "Output" to CargoDependency.smithyRuntimeApi(runtimeConfig).toType().resolve("interceptors::context::Output"),
-            "OutputOrError" to CargoDependency.smithyRuntimeApi(runtimeConfig).toType().resolve("interceptors::context::OutputOrError"),
-            "ResponseDeserializer" to CargoDependency.smithyRuntime(runtimeConfig).toType().resolve("ResponseDeserializer"),
+            "Output" to CargoDependency.smithyRuntimeApi(runtimeConfig).toType().resolve("client::interceptors::context::Output"),
+            "OutputOrError" to CargoDependency.smithyRuntimeApi(runtimeConfig).toType().resolve("client::interceptors::context::OutputOrError"),
+            "ResponseDeserializer" to CargoDependency.smithyRuntime(runtimeConfig).toType().resolve("client::orchestrator::ResponseDeserializer"),
             "SdkBody" to RuntimeType.sdkBody(runtimeConfig),
             "SdkError" to RuntimeType.sdkError(runtimeConfig),
             "TypedBox" to CargoDependency.smithyRuntimeApi(runtimeConfig).toType().resolve("type_erasure::TypedBox"),
