+++ rust-runtime/aws-smithy-runtime/src/client.rs
@@ -1,0 +1,6 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+pub mod orchestrator;
